+++ cookbooks/chub-csdashboard/metadata.rb
@@ -7,7 +7,6 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
 depends 'chub_java'
-depends 'tomcat'
 depends 'openssl'
 depends 'ulimit'
 depends 'mongodb'