+++ cookbooks/chub-onboarder/metadata.rb
@@ -1,4 +1,4 @@
-name             'chub-csdashboard'
+name             'chub-landingstrip'
 maintainer       'CommerceHub Inc.'
 maintainer_email 'estegmann@commercehub.com'
 license          'Apache 2.0'
@@ -6,6 +6,7 @@ description      'Installs/Configures chub-missioncontrol'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends 'java'
-depends 'tomcat'
+depends 'chub_java'
 depends 'openssl'
+depends 'ulimit'
+depends 'mongodb'