+++ cookbooks/chub-onboarder/recipes/default.rb
@@ -17,7 +17,6 @@
 # limitations under the License.
 #
 include_recipe "chub_java::oracle7"
-include_recipe "tomcat"
 
 group "chub-onboarder" do
     action :create
@@ -41,32 +40,14 @@ group "chub-onboarder" do
     members ["chub-onboarder", "chadmin"]
 end
 
-directory node["chub-onboarder"]['app']["config_dir"] do
+directory node["chub-onboarder"]['app']['deploy_dir'] do
   action :create
   owner "chub-onboarder"
   group "chub-onboarder"
   mode 0777
 end
 
-directory "#{node['tomcat']['webapp_dir']}/#{node['chub-onboarder']['app']['app_name']}" do
-  owner "chub-onboarder"
-  group "chub-onboarder"
-  mode 0777
-end
-
-directory "#{node['tomcat']['webapp_dir']}" do
-  owner "chub-onboarder"
-  group "chub-onboarder"
-  mode 0777
-end
-
-directory "#{node['tomcat']['base']}" do
-  owner "chub-onboarder"
-  group "chub-onboarder"
-  mode 0777
-end
-
-directory node["chub-onboarder"]['app']["app_dir"] do
+directory node["chub-onboarder"]['app']["config_dir"] do
   action :create
   owner "chub-onboarder"
   group "chub-onboarder"
@@ -80,25 +61,27 @@ directory node["chub-onboarder"]['app']["log_dir"] do
   mode 0777
 end
 
-execute 'clear_tomcat_app_directory' do
-	command "rm -fr #{node['tomcat']['webapp_dir']}/#{node['chub-onboarder']['app']['app_name']}"
-	action :nothing
-end
-
-file "#{node['tomcat']['webapp_dir']}/#{node['chub-onboarder']['app']['app_name']}.war" do
-	action :delete
-end
-
 touchfile = node['chub-onboarder']['app']['touchfile']
 
-remote_file "#{node['tomcat']['webapp_dir']}/#{node['chub-onboarder']['app']['app_name']}.war" do
-  source "#{node['chub-onboarder']['app']['war_file_url']}"
+remote_file "#{node["chub-onboarder"]['app']['deploy_dir']}/#{node['chub-onboarder']['app']['app_name']}.jar" do
+  source "#{node['chub-onboarder']['app']['jar_file_url']}"
   not_if do
     File.exists?(touchfile)
   end
   owner "chub-onboarder"
   group "chub-onboarder"
   action :create	# This should pull the file down forcefully
-  notifies :run, 'execute[clear_tomcat_app_directory]', :immediately
-  notifies :restart, "service[tomcat]", :delayed
+end
+
+template "/etc/init/onboarder.conf" do
+    source "onboarder.conf.erb"
+    owner "root"
+    group "root"
+    mode 0644
+    notifies "restart", "service[onboarder]", :delayed
+end
+
+service "onboarder" do
+    provider Chef::Provider::Service::Upstart
+    action [ "enable", "start" ]
 end