+++ cookbooks/chub_onboarding_webservice/attributes/default.rb
@@ -1,7 +1,7 @@
 default["chub_onboarding_webservice"]["deployDir"] = "/opt/onboarding_webservice"
 default["chub_onboarding_webservice"]["stagingDir"] = "/opt/staging"
-default["chub_onboarding_webservice"]["jarName"] = "batchagent-deploy.jar"
-default["chub_onboarding_webservice"]["propertyFileName"] = "onboarding.yml"
+default["chub_onboarding_webservice"]["jarName"] = "template-standalone.jar"
+default["chub_onboarding_webservice"]["propertyFileName"] = "config.yml"
 default["chub_onboarding_webservice"]["instanceName"] = "devmcws01"
 default["chub_onboarding_webservice"]["deployJarUrl"] = "OVERRIDE_ME_PLZ"